Moving audio engine to core
